requirements.txt
asgiref==3.3.1
attrs==20.3.0
Automat==20.2.0
cffi==1.14.5
colorama==0.4.4
commonmark==0.9.1
constantly==15.1.0
cryptography==3.4.6
cssselect==1.1.0
dataclasses==0.8
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djongo==1.3.4
dnspython==2.1.0
hyperlink==21.0.0
idna==3.1
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.2
parsel==1.6.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.8.0
PyHamcrest==2.0.2
pymongo==3.11.3
pyOpenSSL==20.0.1
pytz==2021.1
queuelib==1.5.0
rich==9.12.0
Scrapy==2.4.1
scrapy-djangoitem==1.1.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.2.4
Twisted==20.3.0
typing-extensions==3.7.4.3
w3lib==1.22.0
zope.interface==5.2.0
